pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zotero-wordflow"
version = "0.1.0"
description = "Automate DOI verification, Zotero import, and Word docx citation field generation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Lrving-code" }
]
keywords = ["zotero", "word", "docx", "citations", "crossref", "windows"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Text Processing :: Markup :: XML",
  "Topic :: Scientific/Engineering :: Information Analysis"
]
dependencies = [
  "python-docx>=1.1.0",
  "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
zotero-wordflow = "zotero_wordflow.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]